intelligent-data-agent/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
pandas==2.2.1
python-multipart==0.0.9
google-generativeai==0.8.2
python-dotenv==1.0.1
